+++ include/Categories/cat021/cat021_item148.h
@@ -108,9 +108,9 @@ ASTERIX_API uint8_t get_cat021_item148_AM(const cat021_item148 * item);
  * 
  * @param item Pointer to cat021_item148 structure
  * 
- * @return double Altitude in feets
+ * @return int32_t Altitude in feets (-1300ft <= ALT <= 100000 ft)
  */
-ASTERIX_API uint16_t get_cat021_item148_ALT(const cat021_item148 * item);
+ASTERIX_API int32_t get_cat021_item148_ALT(const cat021_item148 * item);
 
 /*******************************************************************************
  * Setters
@@ -148,7 +148,7 @@ ASTERIX_API void set_cat021_item148_AM(cat021_item148 * item, uint8_t src);
  * @param item Pointer to cat021_item148 structure
  * @param value New ALT value (LSB = 25 ft)
  */
-ASTERIX_API void set_cat021_item148_ALT(cat021_item148 * item, double alt);
+ASTERIX_API void set_cat021_item148_ALT(cat021_item148 * item, int32_t alt);
 
 /*******************************************************************************
  * Other Functions
