+++ src/Categories/cat021/cat021_item140.c
@@ -10,7 +10,7 @@
  * Getters
  ******************************************************************************/
 
-ASTERIX_API double get_cat021_item140_GH_raw(const cat021_item140 * item)
+ASTERIX_API uint16_t get_cat021_item140_GH_raw(const cat021_item140 * item)
 {
     return GET_BITS((item)->raw, 1, MASK_16_BITS);
 }
@@ -25,15 +25,14 @@ ASTERIX_API double get_cat021_item140_GH_feet(const cat021_item140 * item)
  ******************************************************************************/
 
 ASTERIX_API void set_cat021_item140_GH_raw(cat021_item140 * item,
-                                            uint16_t gh_raw)
+                                           uint16_t gh_raw)
 {
     // TODO: Check value is within limits
     SET_BITS(&((item)->raw), (gh_raw >> 8), MASK_08_BITS, 9);
     SET_BITS(&((item)->raw), (gh_raw     ), MASK_08_BITS, 1);
 }
 
-ASTERIX_API void set_cat021_item140_GH_feet(cat021_item140 * item,
-                                            double gh)
+ASTERIX_API void set_cat021_item140_GH_feet(cat021_item140 * item, double gh)
 {
     uint16_t gh_raw = 0;
